Adding client-side support for Google deployer
Building Google cli container (gcloud) to run
auth command that generates a OAuth access token
which is needed to be used with any gcloud command.
Saving the generated credentials in google-creds directory
inside LME home directory (~/.lme/). The creds will be
copied into the container that we will build to deploy an
app using the gcloud commands.