AWS - Addressing the case when environment name length is > 40
Beanstalk does not allow environment name to be of length more
than 40 characters. If the name that we generate for the environment
turns out to be more than 40 characters, we truncate it to first 30
characters and add three random digits at the end. We save this
environment name in a file in the application's directory (inside .cld
folder). By using three random digits the probably of collision
with another environment is 1 in 1000 (0.001), which for a given
application might be sufficient. We can always reduce this probability
further by increasing the length of the random string.