AWS - Limiting access to RDS instance to only the application
Earlier the RDS instance was accessible to the world, which was
not good from security point. In this patch we are locking down
the access to only the cidr block corresponding to the default vpc.
This is the virtual private cloud in which the RDS instance is
provisioned by default and in which Elastic Beanstalk
deploys the application by default.