Client-side Google setup update
In order to do Google setup we need to run some gcloud commands
(wrapped in a docker container) from the cld client code.
On Ubuntu we are executing these commands
under 'docker' group as this allows us to not use 'sudo' when
executing these commands. However, this technique does not work
on Mac OS X. On the other hand, it looks like on Mac OS X it
is not necessary to run these commands under 'docker' group.

Therefore, in the client now we are checking for the platform
and executing the google setup commands according to the platform.